requirements.txt
# Minimal dependencies for yamete Python CLI
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
reportlab>=4.0.0
